fix: corrections on customer request
* fix: change directorId to long(BIGINT) type

* fix: make sortBy by enum

* fix: corrected sql query

* fix: corrected sql query №2

* fix: code revision based on comments

* fix: code revision based on comments v2.0

* fix: text of the exceptions into English

* fix: corrected sql query №3

---------

+++ src/main/java/ru/yandex/practicum/filmorate/service/DirectorService.java
@@ -16,7 +16,7 @@ public class DirectorService {
         return directorDao.getAll();
     }
 
-    public Director get(int id) {
+    public Director get(long id) {
         return directorDao.get(id);
     }
 
@@ -29,7 +29,7 @@ public class DirectorService {
         return directorDao.get(director.getId());
     }
 
-    public void delete(int id) {
+    public void delete(long id) {
         directorDao.delete(id);
     }
 }
